Petit changement de la classe main

+++ BarbieRampageGame/main.py
@@ -1,6 +1,5 @@
 # Codé par la CMD-squad
 
-
 import pygame, os, sys
 
 from constants import *
@@ -64,8 +63,8 @@ def spawn_player():
     
     # Création des éléments de l'interface
     player.create_health_bar(10, game_settings.screen_width // 18, assets)
-    player.create_kill_counter(10, game_settings.screen_width * 5/45, assets)
-    player.create_bullet_counter(10, game_settings.screen_width * 33/45, assets)
+    player.create_kill_counter(10, int(game_settings.screen_width * 5/45), assets)
+    player.create_bullet_counter(10, int(game_settings.screen_width * 33/45), assets)
     
     return player
